requirements.txt
# Core dependencies for medical voice agent
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Voice & Audio
deepgram-sdk==3.0.0
pyaudio==0.2.14
wave==0.0.2
websockets==12.0

# AI/ML
huggingface-hub==0.20.1
transformers==4.36.2
torch==2.1.2

# Data processing
numpy==1.24.4
pydantic==2.5.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Utilities
aiofiles==23.2.1
httpx==0.25.2
loguru==0.7.2